requirements.txt
asgiref==3.5.2
Django==3.2.12
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
mysqlclient==2.1.1
PyJWT==1.7.1
pytz==2022.6
sqlparse==0.4.3
